Reorganized Admin search
