+++ app/src/main/java/com/example/mini_pekkas/ui/admin/AdminBrowseFacilities.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.event.admin;
+package com.example.mini_pekkas.ui.admin;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,18 +9,18 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 
-import com.example.mini_pekkas.databinding.FragmentAdminModerationSearchBinding; // Import binding
 import com.example.mini_pekkas.Firebase;
+import com.example.mini_pekkas.databinding.FragmentAdminBrowseFacilitiesBinding;
 
-public class AdminModerationSearchFragment extends Fragment {
+public class AdminBrowseFacilities extends Fragment {
 
-    private FragmentAdminModerationSearchBinding binding; // Declare binding variable
+    private FragmentAdminBrowseFacilitiesBinding binding; // Declare binding variable
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         // Inflate the layout using View Binding
-        binding = FragmentAdminModerationSearchBinding.inflate(inflater, container, false);
+        binding = FragmentAdminBrowseFacilitiesBinding.inflate(inflater, container, false);
         return binding.getRoot(); // Return the root view from binding
     }
