+++ app/src/main/java/com/example/mini_pekkas/ui/admin/AdminBrowseImages.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.event.admin;
+package com.example.mini_pekkas.ui.admin;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,17 +10,19 @@ import androidx.fragment.app.Fragment;
 
 
 import com.example.mini_pekkas.Firebase;
+
 import com.example.mini_pekkas.databinding.FragmentAdminBrowseFacilitiesBinding;
+import com.example.mini_pekkas.databinding.FragmentAdminBrowseImagesBinding;
 
-public class AdminBrowseFacilities extends Fragment {
+public class AdminBrowseImages extends Fragment {
 
-    private FragmentAdminBrowseFacilitiesBinding binding; // Declare binding variable
+    private FragmentAdminBrowseImagesBinding binding; // Declare binding variable
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         // Inflate the layout using View Binding
-        binding = FragmentAdminBrowseFacilitiesBinding.inflate(inflater, container, false);
+        binding = FragmentAdminBrowseImagesBinding.inflate(inflater, container, false);
         return binding.getRoot(); // Return the root view from binding
     }
