+++ app/src/main/java/com/example/mini_pekkas/ui/admin/AdminBrowseProfile.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.event.admin;
+package com.example.mini_pekkas.ui.admin;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,19 +8,21 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.example.mini_pekkas.databinding.FragmentAdminEventDetailsBinding;
-public class AdminEventDetails extends Fragment {
 
-    private FragmentAdminEventDetailsBinding binding; // Declare binding variable
+import com.example.mini_pekkas.databinding.FragmentAdminBrowseProfilesBinding; // Import binding
+import com.example.mini_pekkas.Firebase;
+
+public class AdminBrowseProfile extends Fragment {
+
+    private FragmentAdminBrowseProfilesBinding binding; // Declare binding variable
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         // Inflate the layout using View Binding
-        binding = FragmentAdminEventDetailsBinding.inflate(inflater, container, false);
+        binding = FragmentAdminBrowseProfilesBinding.inflate(inflater, container, false);
         return binding.getRoot(); // Return the root view from binding
     }
-
 
     @Override
     public void onDestroyView() {
