+++ app/src/main/java/com/example/mini_pekkas/ui/admin/AdminEventDetails.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.event.admin;
+package com.example.mini_pekkas.ui.admin;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,23 +8,19 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
+import com.example.mini_pekkas.databinding.FragmentAdminEventDetailsBinding;
+public class AdminEventDetails extends Fragment {
 
-import com.example.mini_pekkas.Firebase;
-
-import com.example.mini_pekkas.databinding.FragmentAdminBrowseFacilitiesBinding;
-import com.example.mini_pekkas.databinding.FragmentAdminBrowseImagesBinding;
-
-public class AdminBrowseImages extends Fragment {
-
-    private FragmentAdminBrowseImagesBinding binding; // Declare binding variable
+    private FragmentAdminEventDetailsBinding binding; // Declare binding variable
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         // Inflate the layout using View Binding
-        binding = FragmentAdminBrowseImagesBinding.inflate(inflater, container, false);
+        binding = FragmentAdminEventDetailsBinding.inflate(inflater, container, false);
         return binding.getRoot(); // Return the root view from binding
     }
+
 
     @Override
     public void onDestroyView() {
