+++ app/src/main/java/com/example/mini_pekkas/ui/admin/AdminModerationSearchFragment.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.event.admin;
+package com.example.mini_pekkas.ui.admin;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,18 +9,18 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 
-import com.example.mini_pekkas.databinding.FragmentAdminBrowseProfilesBinding; // Import binding
+import com.example.mini_pekkas.databinding.FragmentAdminModerationSearchBinding; // Import binding
 import com.example.mini_pekkas.Firebase;
 
-public class AdminBrowseProfile extends Fragment {
+public class AdminModerationSearchFragment extends Fragment {
 
-    private FragmentAdminBrowseProfilesBinding binding; // Declare binding variable
+    private FragmentAdminModerationSearchBinding binding; // Declare binding variable
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         // Inflate the layout using View Binding
-        binding = FragmentAdminBrowseProfilesBinding.inflate(inflater, container, false);
+        binding = FragmentAdminModerationSearchBinding.inflate(inflater, container, false);
         return binding.getRoot(); // Return the root view from binding
     }
