+++ app/src/main/res/layout/fragment_admin_browse_facilities.xml
@@ -23,7 +23,7 @@
 
     <!-- ListView for the list of profiles -->
     <SearchView
-        android:id="@+id/searchView"
+        android:id="@+id/admin_search_facilities"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toBottomOf="@+id/title"
@@ -40,7 +40,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/searchView"
+        app:layout_constraintTop_toBottomOf="@+id/admin_search_facilities"
         app:layout_constraintVertical_bias="1.0" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
