+++ app/src/main/res/navigation/admin_navigation.xml
@@ -7,22 +7,22 @@
 
     <fragment
         android:id="@+id/adminBrowseEvents"
-        android:name="com.example.mini_pekkas.ui.event.admin.AdminBrowseEvent"
+        android:name="com.example.mini_pekkas.ui.admin.AdminBrowseEvent"
         android:label="AdminModerationSearchFragment"
         tools:layout="@layout/fragment_admin_browse_events" />
     <fragment
         android:id="@+id/adminBrowseFacilities"
-        android:name="com.example.mini_pekkas.ui.event.admin.AdminBrowseFacilities"
+        android:name="com.example.mini_pekkas.ui.admin.AdminBrowseFacilities"
         android:label="fragment_admin_event_detail"
         tools:layout="@layout/fragment_admin_browse_facilities" />
     <fragment
         android:id="@+id/adminBrowseImages"
-        android:name="com.example.mini_pekkas.ui.event.admin.AdminBrowseImages"
+        android:name="com.example.mini_pekkas.ui.admin.AdminBrowseImages"
         android:label="AdminModerationSearchFragment"
         tools:layout="@layout/fragment_admin_browse_images" />
     <fragment
         android:id="@+id/adminBrowseProfiles"
-        android:name="com.example.mini_pekkas.ui.event.admin.AdminBrowseProfile"
+        android:name="com.example.mini_pekkas.ui.admin.AdminBrowseProfile"
         android:label="AdminModerationSearchFragment"
         tools:layout="@layout/fragment_admin_browse_profiles" />
 </navigation>
